+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Released on Thursday, November 4 2021.
 
 - Fixed issue with unbalanced grid areas and rows (#82)
 - Fixed small numbers to be transformed into negative exponentials (#80)
+- Fixed issue handling exponential unit values (#79)
 
 # 0.16.1